Fix prerequisites of memcached build
The memcached "GET" script assumed sasl is available and was discovered
by "configure", and then disables it with a patch. If Sasl was not
available the patch would fail. Change this to use "sed", which deletes
the sasl lines regardless of whether it was found or not.

Also added mention in README that libevent is needed.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>