Add "openjdk8" module
Currently, this module fetches OpenJDK's up-to-date sources, and compiles
them, when you do "cd apps/openjdk8; ./GET". This process is *very slow*,
and requires a bunch of unusual development libraries to be installed
(various X-Windows and sound libraries), so if you plan to try this, plan
to spend a lot of time on this compilation.

Now "make image=openjdk8" creates an image with Java 8.

Note, however, that module dependencies are still not done right - if you
do "make image=openjdk8,mgmt.shell", the mgmt module will not know that
openjdk8 is enough, and will bring *also* the "java" module (openjdk7).
This will be fixed in a later patch.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>